// typescript/user-interface/packages/ui-core-components/src/ts/components/ui-widgets/filterable-option-list/types.ts
export interface FilterableOptionListProps {
  options: string[];
  priorityOptions?: string[];
  defaultSelection?: string;
  defaultFilter?: string;
  disabled?: boolean;
  widthPx?: number;
  onSelection(itemSelected: string);
  onEnter?(currentlySelected: string);
  onClick?(itemClicked: string);
  onDoubleClick?(itemClicked: string);
}

export interface FilterableOptionListState {
  currentlySelected: string;
  currentFilter: string;
}